hub/requirements/common.txt
numpy
pillow==8.2.0
boto3
boto3-stubs[essential]
google-auth~=1.30.1
google-cloud-storage~=1.42.0
pathos
humbug>=0.2.6
types-requests
types-click
tqdm